refactor : Emoji 패키지명을 Sticker로 변경 (#21)
