+++ storage/db-core/src/main/kotlin/io/raemian/storage/db/core/sticker/Sticker.kt
@@ -1,4 +1,4 @@
-package io.raemian.storage.db.core.emoji
+package io.raemian.storage.db.core.sticker
 
 import io.raemian.storage.db.core.BaseEntity
 import jakarta.persistence.Column
